Tests for marking a todo as completed
TDD/BDD following a thoughtbot tutorial

A feature test that has a user sign_in, create a todo, then mark that
todo as completed, then check to see if the page has
a .todos li.completed.

Also, a unit test to check the #completed? functionality.
If a todo has a completed_at timestamp, is should "be_completed".
If the timestamp is nil, then it should not "be_completed."