A user can mark a todo as incomplete
This feature was created following TDD/BDD with these specs.

spec/features/user_marks_todo_incomplete_spec.rb
spec/support/features/todo_helper.rb

It modifies the todo index view to use a conditional for choosing which
button (Done or Mark Incomplete) to display. When the Mark Incomplete
button is clicked, the todo instance is routed to completion#destroy
where an `incomplete!` method from the Todo model is invoked. This sets
`completed_at`, which is the basis for determining completion, to `nil`.

This was the last feature built during the tutorial. I hope to continue
working on this app and to continue challenging myself to program using
TDD/BDD.